# openapi.yaml
openapi: 3.1.0
info:
  title: ClassnoteX API
  description: Backend API for ClassnoteX - AI-powered recording and transcription
    app.
  version: 0.4.0
paths:
...
  /share-code/{code}:
    get:
      tags:
      - Share
      summary: Lookup User By Share Code
      operationId: lookup_user_by_share_code_share_code__code__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareCodeLookupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{session_id}/share:
    post:
      tags:
      - Share
      summary: Share Session To User
      operationId: share_session_to_user_sessions__session_id__share_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareByCodeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{session_id}/share/{target_uid}:
    delete:
      tags:
      - Share
      summary: Revoke Share
      operationId: revoke_share_sessions__session_id__share__target_uid__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: target_uid
        in: path
        required: true
        schema:
          type: string
          title: Target Uid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /s/{token}:
    get:
      tags:
      - Share
      summary: Share Fallback
      description: 'Universal Links Fallback Page.

        If the app is installed, the link ''https://deepnote.app/s/{token}'' should
        open the app directly.

        If not installed (or failed), this HTML page is shown.'
      operationId: share_fallback_s__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{session_id}/share_link:
    get:
      tags:
      - Share
      summary: Create Share Link
      operationId: create_share_link_sessions__session_id__share_link_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Share
      summary: Create Share Link
      operationId: create_share_link_sessions__session_id__share_link_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /share/{token}:
    get:
      tags:
      - Share
      summary: Resolve Share Link
      description: 'Share Link Fallback (Proposal B):

        If opened in browser (no app installed), redirect to App Store.

        App with Universal Links will capture this URL and call /share/{token}/info
        instead.'
      operationId: resolve_share_link_share__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /share/{token}/info:
    get:
      tags:
      - Share
      summary: Resolve Share Link Info
      description: Endpoint for the App to resolve token to sessionId.
      operationId: resolve_share_link_info_share__token__info_get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /share/{token}/join:
    post:
      tags:
      - Share
      summary: Join Via Share Link
      operationId: join_via_share_link_share__token__join_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /shares/{token}:
    get:
      tags:
      - Share
      summary: Get Shared Session Details
      description: 'Public read-only session detail view for App (Universal Links).

        No Authentication required (Token is key).'
      operationId: get_shared_session_details_shares__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedSessionDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ShareByCodeRequest:
      properties:
        targetShareCode:
          type: string
          title: Targetsharecode
      type: object
      required:
      - targetShareCode
      title: ShareByCodeRequest
    ShareCodeLookupResponse:
      properties:
        userId:
          type: string
          title: Userid
        displayName:
          anyOf:
          - type: string
          - type: 'null'
          title: Displayname
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
      type: object
      required:
      - userId
      title: ShareCodeLookupResponse
...
    ShareLinkResponse:
      properties:
        url:
          type: string
          title: Url
      type: object
      required:
      - url
      title: ShareLinkResponse
...
    ShareResponse:
      properties:
        sessionId:
          type: string
          title: Sessionid
        sharedUserIds:
          items:
            type: string
          type: array
          title: Shareduserids
      type: object
      required:
      - sessionId
      - sharedUserIds
      title: ShareResponse
    SharedSessionDTO:
      properties:
        sessionId:
          type: string
          title: Sessionid
        title:
          type: string
          title: Title
        transcriptText:
          anyOf:
          - type: string
          - type: 'null'
          title: Transcripttext
        summaryMarkdown:
          anyOf:
          - type: string
          - type: 'null'
          title: Summarymarkdown
        ownerDisplayName:
          anyOf:
          - type: string
          - type: 'null'
          title: Ownerdisplayname
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
      type: object
      required:
      - sessionId
      - title
      title: SharedSessionDTO
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token